Practica2 (#3)
* Add data for practica1
I have added 2 folders. Ecommerce fot django site and etienda for aplication.
In the etienda y have mvt that display querys for the last practise

* Remove .DS_Store from everywhere

* Remove dirty files

* Remove pychache files

* Fix gitignore

* Fix gitignore

* Fix gitignore

* Add static folder in e-commerce
I've added a static folder containing the images folder. This has been done to follow the django convention of putting everything static in this folder

* Add static folder in etienda

* Modify settings
Modify the settings file to add static urls

* Modify Seed file
This file was modified due to relative rute for img folder was changed

* Add cotegories and index templates

* Add request functions

* Modify urls and views files

* practica 2.2 terminada

* Add settings for messages and logs

* Add server logs

* Add products form

* Modify categorias.html and create new product and messages html

* Add data to create product and modify encabezado to improve efficiency

* Add new rules in cards

* Delete img due to encabezado method improvement

* Add new images

* Modify settings for logging

* Modify form
I added some validations for the diferentes fields

* Add sentence to add a new image in a new product

* Add sentence to add a new image in a new product

* Modify template to show validations errors

* Add new viewa for validation error

* Add image librari for the form image

* Fix format

* Add login template

* Add auth features